# SIMDfy >8-bit fjxl on ARM NEON.