Tested on random noise & synthetic images at various bitdepths.

Benchmark on a A51, approximately a 10% speedup on a single core:

```
Before:

./fast_lossless artificial8.ppm x.jxl 2 100
    63.178 MP/s
     2.755 bits/pixel

./fast_lossless artificial13.ppm x.jxl 2 100
    32.670 MP/s
     6.994 bits/pixel

./fast_lossless artificial14.ppm x.jxl 2 100
    32.285 MP/s
     7.956 bits/pixel

./fast_lossless artificial16.ppm x.jxl 2 100
    27.592 MP/s
    10.411 bits/pixel

./fast_lossless bb8.ppm x.jxl 2 20
    37.628 MP/s
    11.895 bits/pixel

./fast_lossless bb13.ppm x.jxl 2 20
    21.407 MP/s
    26.594 bits/pixel

./fast_lossless bb14.ppm x.jxl 2 20
    20.894 MP/s
    29.654 bits/pixel

./fast_lossless bb16.ppm x.jxl 2 20
    18.621 MP/s
    35.989 bits/pixel

After:
./fast_lossless artificial8.ppm x.jxl 2 100
    63.353 MP/s
     2.755 bits/pixel

./fast_lossless artificial13.ppm x.jxl 2 100
    36.205 MP/s
     6.994 bits/pixel

./fast_lossless artificial14.ppm x.jxl 2 100
    34.875 MP/s
     7.956 bits/pixel

./fast_lossless artificial16.ppm x.jxl 2 100
    29.342 MP/s
    10.411 bits/pixel

./fast_lossless bb8.ppm x.jxl 2 20
    38.161 MP/s
    11.895 bits/pixel

./fast_lossless bb13.ppm x.jxl 2 20
    23.457 MP/s
    26.594 bits/pixel

./fast_lossless bb14.ppm x.jxl 2 20
    22.861 MP/s
    29.654 bits/pixel

./fast_lossless bb16.ppm x.jxl 2 20
    20.144 MP/s
    35.989 bits/pixel

```